pyproject.toml
[project]
name = "sql2neo4j"
version = "0.2"
description = "Relational (SQLite) to Neo4j graph mapper"
dependencies = [
    "neo4j>=5.0",
    "mysql-connector-python"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
